# docs/swagger.yaml
definitions:
  v1.errorResponse:
    properties:
      error:
        type: string
    type: object
  v1.getLinkInfoResponse:
    properties:
      dt_created:
        type: string
      dt_will_delete:
        type: string
      long_url:
        type: string
      number_of_clicks:
        type: integer
      short_url:
        type: string
    type: object
  v1.requestMakeShorter:
    properties:
      time_to_live:
        type: integer
      time_to_live_unit:
        type: string
      url:
        type: string
      vip_key:
        type: string
    type: object
  v1.responseMakeShorter:
    properties:
      secret_key:
        type: string
      short_url:
        type: string
    type: object
info:
  contact: {}
  description: Сервис, позволяющий укорачивать ссылки.
  title: Shortener service
  version: "1.0"
paths:
  /{short_url}:
    get:
      description: В строку запроса передаётся краткая ссылка, по ней происходит перенаправление
        на длинную ссылку.
      responses:
        "301":
          description: Moved Permanently
        "404":
          description: Not Found
      summary: Перенаправляет пользователя по ссылке.
  /api/v1/admin/{secret_key}:
    delete:
      description: В строку запроса передаётся секретный ключ, по которому удаляется
        ссылка.
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.errorResponse'
      summary: Удаляет ссылку.
    get:
      description: В строку запроса передаётся секретный ключ, по которому возвращается
        информация о ссылке.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.getLinkInfoResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.errorResponse'
      summary: Возвращает информацию о ссылке.
  /api/v1/make_shorter:
    post:
      description: В формате json передаётся url и опционально остальные параметры.
        В ответ возвращаются секретный ключ и краткая ссылка. Если не передан vip_key,
        то ссылка автоматически генерируется, а если передан, то vip_key и будет использоваться
        как краткий ключ.
      parameters:
      - description: Я не разобрался как тут указать, что всё кроме url nullable,
          поэтому напоминаю! И ещё, если всё-таки делаете vip-key, то обязательными
          становятся все поля
        in: body
        name: make_shorter_request
        required: true
        schema:
          $ref: '#/definitions/v1.requestMakeShorter'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.responseMakeShorter'
        "400":
          description: Сообщения при различных ошибках валидации. Если передано пустое
            поле url, то возвращается ошибка "url is empty". Если передано некорректное
            значение time_to_live_unit или time_to_live <= 0, то возвращается ошибка
            "time to live unit or time to live is invalid". Если vip_key уже сущетсвует
            в базе данных, то возвращается ошибка "vip key is already in use".
          schema:
            $ref: '#/definitions/v1.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.errorResponse'
      summary: Создаёт ссылку.
swagger: "2.0"